requirements.txt
apispec==5.1.1
attrs==21.2.0
awscli==1.20.64
boto3==1.18.64
botocore==1.21.64
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
decorator==5.1.0
docutils==0.15.2
flasgger==0.9.5
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-access-context-manager==0.1.9
google-cloud-asset==3.7.1
google-cloud-container==2.10.1
google-cloud-core==2.2.1
google-cloud-monitoring==2.8.0
google-cloud-org-policy==1.2.1
google-cloud-os-config==1.9.0
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
JSONBender==0.9.3
jsonpath-ng==1.5.3
jsonschema==4.2.1
kubernetes==21.7.0
libcst==0.3.23
MarkupSafe==2.0.1
marshmallow==3.14.1
mistune==2.0.0
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
ply==3.11
proto-plus==1.19.8
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.5.0
six==1.16.0
structlog==21.1.0
toml==0.10.2
typing-extensions==4.0.1
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0

azure-identity
azure-mgmt-resource>=15.0.0
azure-mgmt-compute>=17.0.0
azure-mgmt-monitor>=1.0.0
azure-mgmt-network>=16.0.0
azure-mgmt-storage>=16.0.0
azure-mgmt-containerservice>=14.0.0b1
azure-mgmt-sql>=1.0.0b1

ruamel.yaml